Fix debugLevel persistence and settings save killing streams
- Fix debugLevel not saving: store validated value as int instead of
  float64 so JSON unmarshal to SettingsStruct.DebugLevel (int) works
- Fix settings save killing active streams: only call initBufferVFS()
  when StoreBufferInRAM setting actually changes, not on every save

+++ src/data.go
@@ -126,7 +126,7 @@ func updateServerSettings(request RequestStruct) (settings SettingsStruct, err e
 					if intVal > 3 {
 						intVal = 3
 					}
-					value = float64(intVal)
+					value = intVal
 				}
 
 			}
